[vm] Add an intptr_t cast to fix 32-bit builds
Compilation issue introduced by
commit 0fee95455c8b725991dfd5b3a5d6d017c8895065.

TEST=ci

Cq-Include-Trybots: luci.dart.try:vm-ffi-android-release-arm-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/474401
Commit-Queue: Slava Egorov <vegorov@google.com>

+++ runtime/lib/developer.cc
@@ -247,7 +247,7 @@ DEFINE_NATIVE_ENTRY(Developer_NativeRuntime_streamTimelineTo, 0, 5) {
   if (enable_profiler.value()) {
     Profiler::SetConfig({
         .enabled = true,
-        .period_us = sampling_interval.Value(),
+        .period_us = static_cast<intptr_t>(sampling_interval.Value()),
     });
   }
 #endif
